Add actuator and check GitHub API availability
Adding the Spring Boot Actuator dependency to our application
brings tons of production-ready features for our application.

Looking at `"/actuator"`, we notice a lot of endpoints to
interact with our application and get important information,
including its health (`"/actuator/health"`).

By default, only `"health"` and `"info"` are exposed, and we
can configure our application to expose the ones we want
(see `"management.endpoints.exposure.include"` in
`"application.properties"`).

Because our application depends on the GitHub API, we'd like to
know about its availability. For that, we add a custom
`HealthIndicator` that checks the API and returns useful
information.