Add a custom Micrometer metric for Github API Rate-Limit
Now that we use our own API token to request the GitHub API,
we're less in danger of running out of requests. But we still
have now idea how many requests we have left at any given time.

To get that information, we create a Micrometer Gauge and set
its value using the response headers provided by GitHub.

This way, we can get that instant value using an actuator endpoint.