Enable Caching with Caffeine
Because our Dashboard page is likely to be used a lot, we don't
want to send multiple API requests to GitHub for each page refresh.

First, we need to add the Spring Boot Cache Starter and a cache
implementation - here, Caffeine.
We can now add a new method to our `GithubClient` and stick
a `@Cacheable("events")` annotation on it.

Refreshing the page won't count against our rate-limit if the
cache contains data.